feat: creating Header, renaming files
