+++ src/App.tsx
@@ -1,12 +1,15 @@
 // import { useState } from "react";
 //import FormLogin from "./components/FormLogin/FormLogin";
-import FormRegister from "./components/FormRegister/FormRegister";
+import FormRegister from "./components/FormRegister";
 import "bootstrap/dist/css/bootstrap.min.css";
-import FormLogin from "./components/FormLogin/FormLogin";
+import FormLogin from "./components/FormLogin";
+import Header from "./components/Header";
+
 function App() {
   return (
     <>
-      <FormRegister />
+      {<Header />}
+      {/* <FormRegister /> */}
       {/* <FormLogin/> */}
     </>
   );
