+++ src/components/Header/styles.scss
@@ -1,0 +1,5 @@
+@import "../../assets/styles/variables.scss";
+
+#view-height{
+    height: 9vh;
+}
