+++ src/components/Header/styles.ts
@@ -1,0 +1,6 @@
+import styled from 'styled-components'
+
+export const StyledNav = styled.div`
+  height: 50vh;
+`
+;
